Create First_Order_Error_Propagator
This code produces a tkinter graphical user interface in which a user can input a formula, the values with associated errors and the final value alongside the associated error is outputted. This bases itself on the principle of first-order error propagation. Future iterations may take advantage of higher-order propagation and have a window that prints the inputted formula in a more visual understandable way.